// src/items/enums/item-error-code.enum.ts
export enum ItemErrorCode {
  ITEM_NOT_FOUND = 'ERR_ITEM_NOT_FOUND',
  ITEM_IS_INGREDIENT = 'ERR_ITEM_IS_INGREDIENT', // El error que te salía en texto
  DUPLICATE_SKU = 'ERR_DUPLICATE_SKU',
  DUPLICATE_BARCODE = 'ERR_DUPLICATE_BARCODE',
  DUPLICATE_ENTRY = 'ERR_DUPLICATE_ENTRY', // Genérico por si no detectamos el campo
  LIMIT_REACHED = 'ERR_ITEM_LIMIT_REACHED',
  BULK_LIMIT_EXCEEDED = 'ERR_BULK_LIMIT_EXCEEDED',
  PRO_FEATURE_ONLY = 'ERR_PRO_FEATURE_ONLY',
  BULK_PARTIAL_SUCCESS = 'ERR_BULK_PARTIAL_SUCCESS',
  NAME_EMPTY = 'ERR_NAME_EMPTY',
  INVALID_CONVERSION = 'ERR_INVALID_CONVERSION',
  BASE_UNIT_LOCKED = 'ERR_BASE_UNIT_LOCKED', // Cuando es ingrediente y quieren cambiar la unidad
  INTERNAL_ERROR = 'ERR_INTERNAL_ERROR',
  INSUFFICIENT_STOCK = 'ERR_INSUFFICIENT_STOCK',
  BATCH_LIMIT_EXCEEDED = 'ERR_BATCH_LIMIT_EXCEEDED',
  INVALID_COST_PRICE = 'ERR_INVALID_COST_PRICE',
  INVALID_SALE_PRICE = 'ERR_INVALID_SALE_PRICE',
  INVALID_NUMBER = 'ERR_INVALID_NUMBER',
}